Refactor: searchBar 파일 분리
